break footer routines into their own module
